Write Graph factory function
Graph factory function can be found inside the knightMoves function.
Additionally, I added a newNodes method and a parent property to the
Node factory function. The Graph factory function uses Breadth First
Approach by utilizing a queue and a while loop to create a graph of
moves. Node objects are pushed to the queue, then the next node object
is dequeued in order to create objects from all its possible moves,
which are then enqueued and will be processed in further while loop
iterations. The while loop only finishes iterating when the final
position node is created and processed, which is then what is ultimately
returned.  Each child node (or neighbor) points back to its parent node
through the parent property.